All the functionality about the coding challenge is done

+++ lib/model/src/main/java/com/test/model/weather/WeatherMainInformation.kt
@@ -2,9 +2,9 @@ package com.test.model.weather
 
 data class WeatherMainInformation(
     val temp: Double,
-    val feels_like: Double,
-    val temp_min: Double,
-    val temp_max: Double,
+    val feels_like: Int,
+    val temp_min: Int,
+    val temp_max: Int,
     val pressure: Int,
     val sea_level: Int,
     val grnd_level: Int,
